Fix some circular deps
