+++ js/sdk/src/types/various.ts
@@ -6,65 +6,7 @@
  */
 import { contramap, Ord, ordNumber, ordString } from 'fp-ts/lib/Ord'
 import { Ordering } from 'fp-ts/lib/Ordering'
-import { OffsetMap } from './offsetMap'
-
-/**
- * An Actyx source id.
- * @public
- */
-export type NodeId = string
-const mkNodeId = (text: string): NodeId => text as NodeId
-const randomBase58: (digits: number) => string = (digits: number) => {
-  const base58 = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz'.split('')
-
-  let result = ''
-  let char
-
-  while (result.length < digits) {
-    char = base58[(Math.random() * 57) >> 0]
-    result += char
-  }
-  return result
-}
-
-/**
- * `SourceId` associated functions.
- * @public
- */
-export const NodeId = {
-  /**
-   * Creates a NodeId from a string
-   */
-  of: mkNodeId,
-  /**
-   * Creates a random SourceId with the given number of digits
-   */
-  random: (digits?: number) => mkNodeId(randomBase58(digits || 11)),
-
-  streamNo: (nodeId: NodeId, num: number) => nodeId + '-' + num,
-}
-
-/**
- * An Actyx stream id.
- * @public
- */
-export type StreamId = string
-const mkStreamId = (text: string): StreamId => text as StreamId
-
-/**
- * `SourceId` associated functions.
- * @public
- */
-export const StreamId = {
-  /**
-   * Creates a StreamId from a string
-   */
-  of: mkStreamId,
-  /**
-   * Creates a random StreamId off a random NodeId.
-   */
-  random: () => NodeId.streamNo(mkNodeId(randomBase58(11)), Math.floor(Math.random() * 100)),
-}
+import { NodeId, Offset, OffsetMap, StreamId } from './offsetMap'
 
 /**
  * An Actyx app id.
@@ -94,24 +36,6 @@ const mkLamport = (value: number): Lamport => value as Lamport
 export const Lamport = {
   of: mkLamport,
   zero: mkLamport(0),
-}
-
-/** Offset within an Actyx event stream. @public */
-export type Offset = number
-const mkOffset = (n: number): Offset => n as Offset
-
-/** Functions related to Offsets. @public */
-export const Offset = {
-  of: mkOffset,
-  zero: mkOffset(0),
-  /**
-   * A value that is below any valid Offset
-   */
-  min: mkOffset(-1),
-  /**
-   * A value that is above any valid Offset
-   */
-  max: mkOffset(Number.MAX_SAFE_INTEGER),
 }
 
 /** Timestamp (UNIX epoch), MICROseconds resolution. @public */
